Add Default Auth Middleware & add polymoph relation

+++ bin/controller.js
@@ -1,6 +1,7 @@
 import fs from "fs";
 import path from "path";
 import { baseEnv } from "devlien/env";
+import DIR from "../framework/App/Core/Helpers/DIR.js";
 
 export default class Controller {
 
@@ -18,11 +19,21 @@ export default class Controller {
 
         let file = `app/Http/Controllers/${controllerName}.js`
 
+        const dir = DIR.path(path.join(baseEnv.BASE_PATH, file));
+
+        if(dir.isExist()){
+            console.warn('Controller Already Exists');
+            return true;
+        }
+        dir.make();
+
         terminal.addLine(`${file} @space GENERATING`);
 
         let mgn = new Controller;
         var content = fs.readFileSync(mgn.package_path+'/libraries/standard/controller.js', 'utf-8');
-        content = content.replaceAll('@controller', controllerName)
+        
+        content = content.replaceAll('@controller', dir.filename(false));
+
         fs.writeFileSync(path.join(baseEnv.BASE_PATH, file), content);
 
         terminal.addLine(`${file} @space GENERATED`, 'success');
